I have made some changes to the the UI in the mobile application. And also want to implement the genres to load the available genres.

+++ app/src/main/java/com/dennismoviedb/moviedb/model/Movie.java
@@ -10,7 +10,7 @@ import org.parceler.Parcel;
 public class Movie {
     private String movie_title;
     private String movie_poster;
-    private int movie_vote;
+    private double movie_vote;
     private String movie_backdropImage;
     private String movie_overview;
     private boolean adult;
@@ -18,7 +18,7 @@ public class Movie {
     private int id;
 
 
-    public Movie(String movie_title, String movie_poster, int movie_vote,
+    public Movie(String movie_title, String movie_poster, double movie_vote,
                  String movie_backdropImage, String movie_overview, boolean adult, String pageNumber, int id){
         this.movie_title = movie_title;
         this.movie_poster = movie_poster;
@@ -40,7 +40,7 @@ public class Movie {
         return movie_poster;
     }
 
-    public int getMovie_vote() {
+    public double getMovie_vote() {
         return movie_vote;
     }
